+++ include/Account.h
@@ -4,10 +4,11 @@
 
 #ifndef LAB_04_BOOST_FILESYSTEM_ACCOUNT_H
 #define LAB_04_BOOST_FILESYSTEM_ACCOUNT_H
-
+#pragma once
 #include <string>
 class Account {
  public:
+  Account(const std::string& a,const std::string& b, size_t c,const std::string& d);
   std::string broker;
   std::string num_account;
   size_t amount_of_files;
