--- sources/source.cpp
@@ -1,3 +1,0 @@
-// Copyright 2018 Your Name <your_email>
-
-#include <header.hpp>
