+++ tests/test_runner.cpp
@@ -1,0 +1,14 @@
+//Copyright [2020] <Lev Semkov>
+//
+// Created by levs503 on 28.10.2020.
+//
+
+
+#include <gtest/gtest.h>
+
+int main(int argc, char * argv[]) {
+  testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
+
+
